+++ server/lib/related-content/story-package.js
@@ -2,14 +2,13 @@
 const getArticle = require('../get-article');
 const dateTransform = require('../article-date');
 const sanitizeImage = require('../sanitize-image');
-const getStreamUrl = require('../get-stream-url');
 const url = require('../url');
 
 const formatRelatedContent = (options, item) => {
 	const primaryTheme = (item.metadata || []).filter(metadatum => !!metadatum.primary)[0];
 	options._wrappedFetchGroup = `story-package-${item.id}`;
 
-	return getStreamUrl(primaryTheme, options)
+	return url.stream(primaryTheme, options)
 		// Ignore errors
 		.catch(() => {})
 		.then(streamUrl => ({
